chore(editor): Drop `UPLOAD_LABEL_REBUILD` flag from public interface
